FE: Move directory of json data files
