+++ frontend/src/data/json/items/fonts.json
