+++ frontend/src/data/json/items/papers.json
