+++ frontend/src/data/json/items/stickers.json
